[refactor] ItemDropController에서 ItemController로 패키지명 및 클래스명 리팩토링
