+++ src/main/java/com/depromeet/streetdrop/apis/item/controller/ItemController.java
@@ -1,8 +1,8 @@
-package com.depromeet.streetdrop.apis.item.drop.controller;
+package com.depromeet.streetdrop.apis.item.controller;
 
 import com.depromeet.streetdrop.domains.common.dto.ResponseDto;
-import com.depromeet.streetdrop.domains.item.drop.dto.ItemDropRequestDto;
-import com.depromeet.streetdrop.domains.item.drop.service.ItemDropService;
+import com.depromeet.streetdrop.domains.item.dto.ItemRequestDto;
+import com.depromeet.streetdrop.domains.item.service.ItemService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -13,16 +13,16 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping(value = "/items", produces = "application/json;charset=UTF-8")
 @RequiredArgsConstructor
 @Tag(name = "Items", description = "Item Drop API")
-public class ItemDropController {
-	private final ItemDropService itemDropService;
+public class ItemController {
+	private final ItemService itemService;
 
 	@Operation(summary = "드랍 아이템 등록")
 	@PostMapping("/{memberId}")
 	public ResponseEntity<Void> create(
 			@PathVariable Long memberId,
-			@RequestBody ItemDropRequestDto requestDto
+			@RequestBody ItemRequestDto requestDto
 	) {
-		itemDropService.register(memberId, requestDto);
+		itemService.register(memberId, requestDto);
 		return ResponseDto.noContent();
 	}
 }
