+++ src/main/java/com/depromeet/streetdrop/domains/item/repository/ItemRepository.java
@@ -1,0 +1,9 @@
+package com.depromeet.streetdrop.domains.item.repository;
+
+import com.depromeet.streetdrop.domains.item.entity.Item;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ItemRepository extends JpaRepository<Item, Long> {
+}
