+++ src/main/java/com/depromeet/streetdrop/domains/itemLocation/entity/ItemLocation.java
@@ -2,7 +2,7 @@ package com.depromeet.streetdrop.domains.itemLocation.entity;
 
 import com.depromeet.streetdrop.domains.area.village.entity.VillageArea;
 import com.depromeet.streetdrop.domains.common.BaseTimeEntity;
-import com.depromeet.streetdrop.domains.item.drop.entity.Item;
+import com.depromeet.streetdrop.domains.item.entity.Item;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
