+++ src/main/java/com/depromeet/streetdrop/domains/user/entity/User.java
@@ -1,7 +1,7 @@
 package com.depromeet.streetdrop.domains.user.entity;
 
 import com.depromeet.streetdrop.domains.common.BaseTimeEntity;
-import com.depromeet.streetdrop.domains.item.drop.entity.Item;
+import com.depromeet.streetdrop.domains.item.entity.Item;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
